+++ src/library/modal.js
@@ -1,8 +1,5 @@
-import { setup, Modal } from './controllers/modal';
+import { setup } from './controllers/modal';
 import { addPerformanceMeasure } from '../utils';
-
-window.paypal = window.paypal ?? {};
-window.paypal.MessagesModal = Modal;
 
 addPerformanceMeasure('scriptLoadDelay');
 setup();