+++ src/library/zoid/modal/component.js
@@ -164,9 +164,9 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                             index,
                             refIndex,
                             et: 'CLICK',
-                            event_type: 'click',
-                            link: linkName,
-                            src: src ?? linkName
+                            event_type: 'modal_rendered',
+                            page_view_link_name: linkName,
+                            page_view_link_source: src ?? linkName
                         });
 
                         if (typeof onClick === 'function') {
@@ -192,14 +192,37 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                             index,
                             refIndex,
                             et: 'CLICK',
-                            event_type: 'click',
-                            link: 'Calculator',
-                            src: 'Calculator',
-                            amount: value
+                            event_type: 'modal_rendered',
+                            page_view_link_name: 'Calculator',
+                            page_view_link_source: 'Calculator',
+                            calculator_input: value
                         });
 
                         if (typeof onCalculate === 'function') {
                             onCalculate({ value });
+                        }
+                    };
+                }
+            },
+            onError: {
+                type: 'function',
+                queryParam: false,
+                value: ({ props }) => {
+                    const { onError } = props;
+
+                    return ({ message }) => {
+                        const { index, refIndex } = props;
+
+                        logger.track({
+                            index,
+                            refIndex,
+                            et: 'ERROR',
+                            event_type: 'error',
+                            message
+                        });
+
+                        if (typeof onError === 'function') {
+                            onError({ message });
                         }
                     };
                 }
@@ -217,7 +240,7 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                             index,
                             refIndex,
                             et: 'CLIENT_IMPRESSION',
-                            event_type: 'modal-open',
+                            event_type: 'modal_viewed',
                             src
                         });
 
@@ -247,8 +270,8 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                             index,
                             refIndex,
                             et: 'CLICK',
-                            event_type: 'modal-close',
-                            link: linkName
+                            event_type: 'modal_close',
+                            page_view_link_name: linkName
                         });
 
                         if (typeof onClose === 'function') {
@@ -314,7 +337,7 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                             index,
                             refIndex,
                             et: 'CLIENT_IMPRESSION',
-                            event_type: 'modal-render',
+                            event_type: 'modal_rendered',
                             modal: `${products.join('_').toLowerCase()}:${offer ? offer.toLowerCase() : products[0]}`,
                             // For standalone modal the stats event does not run, so we duplicate some data here
                             bn_code: getScriptAttributes()[SDK_SETTINGS.PARTNER_ATTRIBUTION_ID],
@@ -449,7 +472,7 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                 type: 'string',
                 queryParam: 'features',
                 required: false,
-                value: validate.features ?? getFeatures
+                value: ({ props }) => getFeatures(validate.features({ props }))
             }
         }
     })