+++ src/server/locale/US/CONTEXTUAL_COMPONENTS/MARKS/mutations/long_term_multi_gtz.js
@@ -1,0 +1,26 @@
+import { xSmallFallback, hideDisclaimer } from '../../../../../message/mediaQueries';
+
+export default {
+    'layout:text': [
+        [
+            'default',
+            ({ textSize }) => ({
+                styles: [
+                    `.message__headline > .tag--xsmall > span:first-child { white-space: normal; } `,
+                    xSmallFallback(textSize * 14),
+                    hideDisclaimer(textSize * 14)
+                ],
+                headline: [
+                    {
+                        tag: 'medium'
+                    },
+                    {
+                        tag: 'xsmall',
+                        br: ['or', ',']
+                    }
+                ],
+                disclaimer: ['default']
+            })
+        ]
+    ]
+};