+++ tests/functional/v2/utils/collectDiffs.js
@@ -77,7 +77,9 @@ const uploadToImgur = async subDir => {
     await collectDiffs();
 
     const [bannerDiffs, modalDiffs] = await Promise.all([uploadToImgur('modal'), uploadToImgur('banner')]).catch(
-        error => console.error(error) // eslint-disable-line no-console
+        error => {
+            console.error(error); // eslint-disable-line no-console
+        }
     );
 
     if (bannerDiffs + modalDiffs > 0) {