+++ tests/unit/spec/src/components/modal/v2/lib/zoid-polyfill.test.js
@@ -108,7 +108,7 @@ describe('zoidPollyfill', () => {
             expect(logger.track).toHaveBeenCalledTimes(1);
             expect(logger.track).toHaveBeenCalledWith(
                 expect.objectContaining({
-                    event_type: 'modal-render',
+                    event_type: 'modal_rendered',
                     modal: 'product_1_product_2:PRODUCT_1'
                 })
             );
@@ -119,9 +119,9 @@ describe('zoidPollyfill', () => {
             expect(logger.track).toHaveBeenCalledTimes(1);
             expect(logger.track).toHaveBeenCalledWith(
                 expect.objectContaining({
-                    event_type: 'click',
-                    link: 'test link',
-                    src: 'test src'
+                    event_type: 'modal_rendered',
+                    page_view_link_name: 'test link',
+                    page_view_link_source: 'test src'
                 })
             );
         });
@@ -131,10 +131,10 @@ describe('zoidPollyfill', () => {
             expect(logger.track).toHaveBeenCalledTimes(1);
             expect(logger.track).toHaveBeenCalledWith(
                 expect.objectContaining({
-                    event_type: 'click',
-                    link: 'Calculator',
-                    src: 'Calculator',
-                    amount: '500'
+                    event_type: 'modal_rendered',
+                    page_view_link_name: 'Calculator',
+                    page_view_link_source: 'Calculator',
+                    calculator_input: '500'
                 })
             );
         });
@@ -144,8 +144,8 @@ describe('zoidPollyfill', () => {
             expect(logger.track).toHaveBeenCalledTimes(1);
             expect(logger.track).toHaveBeenCalledWith(
                 expect.objectContaining({
-                    event_type: 'modal-open',
-                    src: 'Show'
+                    event_type: 'modal_viewed',
+                    page_view_link_source: 'Show'
                 })
             );
         });
@@ -155,8 +155,8 @@ describe('zoidPollyfill', () => {
             expect(logger.track).toHaveBeenCalledTimes(1);
             expect(logger.track).toHaveBeenCalledWith(
                 expect.objectContaining({
-                    event_type: 'modal-close',
-                    link: 'Close Button'
+                    event_type: 'modal_close',
+                    page_view_link_name: 'Close Button'
                 })
             );
         });
@@ -168,8 +168,8 @@ describe('zoidPollyfill', () => {
                 expect.objectContaining({
                     index: '1',
                     et: 'CLICK',
-                    event_type: 'modal-close',
-                    link: 'Escape Key'
+                    event_type: 'modal_close',
+                    page_view_link_name: 'Escape Key'
                 })
             );
         });